+++ src/main/java/com/masuda/tabletopmanager/model/Personagem/Resumo/TftPersonagemResumo.java
@@ -3,14 +3,14 @@ package com.masuda.tabletopmanager.model.Personagem.Resumo;
 import java.util.Map;
 
 public record TftPersonagemResumo(
-    long id,
+    Integer id,
     String nome,
     String sistema,
     String avatarUrl,
     String sin
 ) implements PersonagemResumo {
     public static TftPersonagemResumo converterRegistros(Map<String, Object> registros) {
-    long id = (Long) registros.get("id");
+    Integer id = (Integer) registros.get("id");
     String nome = (String) registros.get("nome");
     String sistema = (String) registros.get("sistema");
     String avatarUrl = (String) registros.get("avatar_url");
